modify:goqZipCodeを外部ファイルへ切り出しました
